.flow/pyproject.toml
[project]
name = "ai-flow"
version = "0.1.0"
description = "Recursive task scheduling + multi-Agent collaboration framework on GitHub"
requires-python = ">=3.11"
dependencies = [
    "ruamel.yaml>=0.18.0",
    "gitpython>=3.1",
    "pygithub>=2.1",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
flow = "flow.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
flow = ["bootstrap/**/*", "bootstrap/**/.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
known-first-party = ["flow"]
